+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.21.10
+
+- Fix issue in the computation of `pubHostedUrl` in `computeOutdatedVersions`.
 ## 0.21.9
 
 - New defaults for JSON: do not include `null` values, explicit `toJson()`.